MAGMA 2.0 beta 2016-01-06
MAGMA 2.0 beta has been released. This includes a major interface change
for all MAGMA BLAS functions; most higher level functions such as
magma_zgetrf have not changed their interface. Significant changes:

 - Added queue argument to magmablas routines, and deprecated
   magmablas{Set,Get}KernelStream. This resolves a thread safety issue
   with using global magmablas{Set,Get}KernelStream.
 - Fixed bugs related to relying on CUDA NULL stream implicit
   synchronization.
 - Fixed memory leaks (zunmqr_m, zheevdx_2stage, etc.). Add
  -DDEBUG_MEMORY option to catch leaks.
 - Fixed geqrf*_gpu bugs for m == nb, n >> m (ex: -N 64,10000); and m >>
   n, n == nb+i (ex: -N 10000,129)
 - Fixed zunmql2_gpu for rectangular sizes.
 - Fixed zhegvdx_m itype 3.
 - Added zunglq, zungbr, zgeadd2 (which takes both alpha and beta).

MAGMA sparse
 - Added QMR, TFQMR, preconditioned TFQMR
 - Added CGS, preconditioned CGS
 - Added kernel-fused versions for CGS/PCGS QMR, TFQMR/PTFQMR
 - Changed relative stopping criterion to be relative to RHS
 - Fixed bug in complex version of CG
 - Accelerated version of Jacobi-CG
 - Added very efficient IDR
 - Performance tuning for SELLP SpMV

The Windows port using CMake has not yet been implemented in the new
Makefile structure. It will be forthcoming in the final release.

+++ control/abs.cpp
@@ -34,8 +34,8 @@ magma_cabs(magmaDoubleComplex z)
 extern "C" float
 magma_cabsf(magmaFloatComplex z)
 {
-    float x = fabs( real( z ));
-    float y = fabs( imag( z ));
+    float x = fabsf( real( z ));
+    float y = fabsf( imag( z ));
     float big, little;
     if ( x > y ) {
         big    = x;
